+++ server/graphql/base/resolvers.ts
@@ -7,6 +7,12 @@ import botApproveResolver from "../bot/resolvers/bot.approve.resolver";
 import userGetResolver from "../user/resolvers/user.get.resolver";
 import userUpdateResolver from "../user/resolvers/user.update.resolver";
 import userDelete from "../user/resolvers/user.delete.resolver";
+
+import reviewGetResolver from "../review/resolvers/review.get.resolver";
+import reviewCreateResolver from "../review/resolvers/review.create.resolver";
+import reviewDeleteResolver from "../review/resolvers/review.delete.resolver";
+import reviewLikeResolver from "../review/resolvers/review.like.resolver";
+import reviewDislikeResolver from "../review/resolvers/review.dislike.resolver";
 // import reviewResolver from "../bot/resolvers/bot.create.resolver";
 // import reviewMutationResolver from "../review/review.mutation.resolver";
 // import userResolver from "../user/user.resolver";
@@ -24,6 +30,13 @@ const resolvers = [
     userGetResolver,
     userUpdateResolver,
     userDelete,
+
+    reviewGetResolver,
+    reviewCreateResolver,
+    reviewDeleteResolver,
+    reviewLikeResolver,
+    reviewDislikeResolver,
+
 
     // reviewResolver,
     // reviewMutationResolver,
