+++ server/graphql/base/typeDefs.ts
@@ -1,5 +1,4 @@
 import query from "./queries";
-import mutations from "./mutations";
 import botType from "../bot/bot.type";
 import reviewType from "../review/review.type";
 import ownerReplyType from "../ownerReply/ownerReply.type";
@@ -7,6 +6,13 @@ import userType from "../user/user.type";
 import notificationType from "../notification/notification.type";
 import voteType from "../vote/vote.type";
 
-const typeDefs = [query, mutations, botType, reviewType, ownerReplyType, userType, notificationType, voteType];
+const typeDefs = [
+    query,
+    botType,
+    reviewType,
+    ownerReplyType,
+    userType,
+    notificationType,
+    voteType];
 
 export default typeDefs;
